feat(crypto): wire the local catalog into the discovery resolver
makeMarketDataServices now opens a read-only lookup handle to the shared
CoinGecko catalog (no refresh task — the registry-wiring catalog owns that;
WAL lets this connection see the refreshed snapshot) and injects it as the
discovery resolver's localResolver. A known token (canonical USDC, any
catalog-listed token) now resolves to .priced offline, instead of sitting
.unpriced behind a rate-limited anonymous CoinGecko call.

Gated off under UI testing via the non-MainActor currentUITestSeed(), so the
DI ordering and @MainActor isolation of the app-boot path are untouched.
Catalog helpers moved to ProfileSession+CatalogFactory.swift to keep
ProfileSession+Factories.swift under the file-length limit.
